+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# [1.4]
+
+- **New feature** - eslint-loader removed from webpack compile.
+
+# [1.3]
+
+- **New feature** - `mope tx status` to get list of transifex resources and
+  their completion status.
+
+
 # [1.2]
 
 - **New feature** - `mope gh status` to set PR status. Run `mope gh status -h`